requirements.txt
# Google Drive API
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Core vision & ML
opencv-python>=4.9.0
Pillow>=10.3.0
numpy>=1.26.0
torch>=2.2.0
torchvision>=0.17.0

# Data handling
pandas>=2.2.0
tqdm>=4.66.0

# Feature extraction (classical)
scikit-image>=0.22.0

# Feature extraction (deep)
# timm covers DINOv2, EfficientNet, ViT, etc.
timm>=0.9.16

# Approximate nearest neighbour search (fast retrieval)
faiss-cpu>=1.8.0        # swap for faiss-gpu if CUDA available

# Evaluation metrics
scikit-learn>=1.4.0

# Visualisation & notebooks
matplotlib>=3.8.0
seaborn>=0.13.0
ipykernel>=6.29.0
jupyter>=1.0.0
